Module icu_pluralrules::data

Temporary module with logic required to provide data to the crate.

Examples

use icu_locale::LanguageIdentifier;
use icu_pluralrules::{PluralCategory, PluralRuleType, PluralOperands};

use icu_pluralrules::data::provider::{DummyDataProvider, DataProviderType};

let dp = DummyDataProvider::default();

// Dummy provider always returns data for "en"
let lang: LanguageIdentifier = "en".parse()
    .expect("Failed to parse a language identifier.");

// Dummy provider always returns data for cardinal
let cardinal_type = PluralRuleType::Cardinal;

let selector = dp.get_selector(&lang, cardinal_type)
    .expect("Failed to retrieve a rule.")
    .expect("Requested rule set does not exist.");

let operands: PluralOperands = 1_usize.into();

assert_eq!(selector.select(&operands), PluralCategory::One);

Modules

cldr_resource

io

provider
Traits for Data Provider and a dummy Data Provider

Enums

RulesSelector
An enum storing models of handling plural rules selection.

Type Definitions

PluralRuleList
A list of tuples of (PluralCategory-ast::Condition) pairs.

PluralRulesFn
A raw function pointer to a PluralRulesFn