Module icu_pluralrules::data::provider

Traits for Data Provider and a dummy Data Provider

Every data provider must implement the DataProviderType.

For the example below, we're going to implement a simple data provider for an imaginary language which has three Plural Categories: One, Few, and Other.

The provider takes a syntax described in rules module and parses it into a vector of category-condition tuples.

Such provider can be then passed to the PluralRules struct.

Examples

use icu_pluralrules::{PluralCategory, PluralRuleType};
use icu_pluralrules::data::RulesSelector;
use icu_pluralrules::data::provider::{DataProviderType, DataProviderError};
use icu_pluralrules::rules::ast;
use icu_locale::LanguageIdentifier;

use icu_pluralrules::rules::parse_condition;

struct MyDataProvider {}

impl DataProviderType for MyDataProvider {
    fn get_selector(
        &self,
        locale: &LanguageIdentifier,
        type_: PluralRuleType,
    ) -> Result<Option<RulesSelector>, DataProviderError> {

        let sources = vec![
            (PluralCategory::One, "i = 1 and v = 0"),
            (PluralCategory::Few, "i = 2..4 and v != 0"),
            (PluralCategory::Other, "")
        ];

        let conditions: Vec<(PluralCategory, ast::Condition)> =
            sources.iter().map(|(category, rule_str)| {
                let condition = parse_condition(rule_str.as_bytes())
                    .expect("Failed to parse the plural rule.");
                (*category, condition)
            }).collect();

        Ok(
            Some(
                RulesSelector::Conditions(
                    conditions.into_boxed_slice()
                )
            )
        )
    }
}

Structs

DummyDataProvider
A dummy implementation of the DataProviderType for use in tests and examples.

Enums

DataProviderError
A list of possible error outcomes for the DataProviderType trait.

Traits

DataProviderType
Trait required to be implemented by any Data Provider.