Trait icu_pluralrules::data::provider::DataProviderType

pub trait DataProviderType {
    fn get_selector(
        &self, 
        locale: &LanguageIdentifier, 
        type_: PluralRuleType
    ) -> Result<Option<RulesSelector>, DataProviderError>;
}

Trait required to be implemented by any Data Provider.

See provider module docs for example on how to implement the trait.

Examples

use icu_locale::LanguageIdentifier;
use icu_pluralrules::{PluralCategory, PluralRuleType, PluralOperands};

use icu_pluralrules::data::provider::{DummyDataProvider, DataProviderType};

let dp = DummyDataProvider::default();

// Dummy provider always returns data for "en"
let lang: LanguageIdentifier = "en".parse()
    .expect("Failed to parse a language identifier.");

// Dummy provider always returns data for cardinal
let cardinal_type = PluralRuleType::Cardinal;

let selector = dp.get_selector(&lang, cardinal_type)
    .expect("Failed to retrieve a rule.")
    .expect("Requested rule set does not exist.");

let operands: PluralOperands = 1_usize.into();

assert_eq!(selector.select(&operands), PluralCategory::One);

Required methods

fn get_selector(
&self,
locale: &LanguageIdentifier,
type_: PluralRuleType
) -> Result<Option<RulesSelector>, DataProviderError>

Returns a RulesSelector variant for a requested locale and type.

Loading content...